refactor: OrderTableRequest 패키지 변경
